:test_tube:  testcode 작성을 위한 의존성 주입 및 PostController test
- PostController: RetrievePostResponse가 null을 리턴하지 않도록 postService.findById(postId)가 null을 반환할 경우 NOT_FOUND 상태와 함께 null을 반환하도록 수정
- PostControllerTest: BehaviorSpec을 사용하여 응답 코드와 ID가 올바르게 저장되었는지 확인하는 테스트 코드 추가 (프론트엔드 및 클라이언트와의 소통을 위해)

+++ src/main/kotlin/com/teamsparta/assignment/domain/post/model/Post.kt
@@ -32,7 +32,7 @@ class Post (
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Int? = null
+    var id: Long? = null
 
     @OneToMany(mappedBy = "post", fetch = FetchType.LAZY, cascade = [CascadeType.REMOVE])
     val comments : List<Comment> = emptyList()
